backend/requirements.txt
fastapi
pandas
numpy
scipy
scikit-learn
python-multipart
uvicorn
liac-arff
tensorflow
